@@ -1,7 +1,7 @@
 [workspace.package]
-version = "1.1.2"
+version = "1.3.0"
 edition = "2024"
-rust-version = "1.91.0"
+rust-version = "1.93.0"
 license = "MIT OR Apache-2.0"
 publish = false
 
