@@ -148,6 +148,14 @@ pub struct MaxTpsArgs {
     #[arg(long, default_value_t = 0.0)]
     erc20_weight: f64,
 
+    /// Send transfers to existing signer accounts instead of random new addresses.
+    ///
+    /// When enabled, TIP-20 and ERC-20 transfers are sent to the bench's own signer addresses
+    /// (which already exist on-chain), avoiding cold SSTORE for account creation. This tests
+    /// pure transfer throughput without state growth.
+    #[arg(long)]
+    existing_recipients: bool,
+
     /// An amount of receipts to wait for after sending all the transactions.
     #[arg(long, default_value_t = 100)]
     sample_size: usize,
@@ -375,6 +383,20 @@ impl MaxTpsArgs {
         let swap_weight = (self.swap_weight * Self::WEIGHT_PRECISION).trunc() as u64;
         let erc20_weight = (self.erc20_weight * Self::WEIGHT_PRECISION).trunc() as u64;
 
+        let recipients = if self.existing_recipients {
+            let addrs: Vec<Address> = signer_providers
+                .iter()
+                .map(|(signer, _)| signer.address())
+                .collect();
+            info!(
+                recipients = addrs.len(),
+                "Using existing signer addresses as recipients"
+            );
+            Some(addrs)
+        } else {
+            None
+        };
+
         let gen_input = GenerateTransactionsInput {
             total_txs,
             accounts,
@@ -387,6 +409,8 @@ impl MaxTpsArgs {
             quote_token,
             user_tokens,
             erc20_tokens,
+            recipients,
+            tx_id: Arc::new(AtomicUsize::new(0)),
         };
 
         // For expiring nonces, we need to generate/sign/send in batches to avoid
@@ -742,6 +766,8 @@ async fn generate_transactions<F: TxFiller<TempoNetwork> + 'static>(
         quote_token,
         user_tokens,
         erc20_tokens,
+        recipients,
+        tx_id,
     } = input;
 
     let txs_per_sender = total_txs / accounts;
@@ -763,6 +789,8 @@ async fn generate_transactions<F: TxFiller<TempoNetwork> + 'static>(
     let swaps = Arc::new(AtomicUsize::new(0));
     let orders = Arc::new(AtomicUsize::new(0));
     let erc20_transfers = Arc::new(AtomicUsize::new(0));
+    // `tx_id` is shared across batches via `GenerateTransactionsInput` to ensure
+    // unique tx hashes across all batches when using expiring nonces.
 
     let builders = ProgressBar::new(total_txs)
         .wrap_stream(stream::iter(
@@ -780,6 +808,11 @@ async fn generate_transactions<F: TxFiller<TempoNetwork> + 'static>(
                 .choose_weighted(&mut rand::rng(), |(_, weight)| *weight)?
                 .0;
 
+            let recipient = match &recipients {
+                Some(addrs) => *addrs.choose(&mut rand::rng()).unwrap(),
+                None => Address::random(),
+            };
+
             let mut tx = match tx_index {
                 0 => {
                     tip20_transfers.fetch_add(1, Ordering::Relaxed);
@@ -787,7 +820,7 @@ async fn generate_transactions<F: TxFiller<TempoNetwork> + 'static>(
 
                     // Transfer minimum possible amount
                     token
-                        .transfer(Address::random(), U256::ONE)
+                        .transfer(recipient, U256::ONE)
                         .into_transaction_request()
                 }
                 1 => {
@@ -820,7 +853,7 @@ async fn generate_transactions<F: TxFiller<TempoNetwork> + 'static>(
 
                     // Transfer minimum possible amount
                     token
-                        .transfer(Address::random(), U256::ONE)
+                        .transfer(recipient, U256::ONE)
                         .into_transaction_request()
                 }
                 _ => unreachable!("Only {TX_TYPES} transaction types are supported"),
@@ -870,7 +903,19 @@ async fn generate_transactions<F: TxFiller<TempoNetwork> + 'static>(
                 });
             }
 
-            eyre::Ok((tx.try_into_request()?, signer))
+            let mut req = tx.try_into_request()?;
+
+            // Bump priority fee by a unique counter to ensure unique tx hashes
+            // when using expiring nonces (which share nonce=0).
+            let id = tx_id.fetch_add(1, Ordering::Relaxed) as u128;
+            if let Some(fee) = req.max_priority_fee_per_gas() {
+                req.inner.set_max_priority_fee_per_gas(fee + id);
+            }
+            if let Some(fee) = req.max_fee_per_gas() {
+                req.inner.set_max_fee_per_gas(fee + id);
+            }
+
+            eyre::Ok((req, signer))
         })
         .buffer_unordered(max_concurrent_requests)
         .try_collect::<Vec<_>>()
@@ -1163,4 +1208,9 @@ struct GenerateTransactionsInput<F: TxFiller<TempoNetwork>> {
     quote_token: Address,
     user_tokens: Vec<Address>,
     erc20_tokens: Vec<Address>,
+    /// When set, transfers go to these existing addresses instead of `Address::random()`.
+    recipients: Option<Vec<Address>>,
+    /// Shared counter used to bump priority fee, ensuring unique tx hashes across batches
+    /// when using expiring nonces (which share nonce=0).
+    tx_id: Arc<AtomicUsize>,
 }
