@@ -21,11 +21,21 @@ use reth_network_peers::NodeRecord;
 use std::sync::{Arc, LazyLock};
 use tempo_primitives::TempoHeader;
 
-/// T0 base fee: 10 gwei (1×10^10 wei)
+/// T0 base fee: 10 billion attodollars (1×10^10)
+///
+/// Attodollars are the atomic gas accounting units at 10^-18 USD precision.
+/// Basefee is denominated in attodollars.
 pub const TEMPO_T0_BASE_FEE: u64 = 10_000_000_000;
 
-/// T1 base fee: 20 gwei (2×10^10 wei)
-/// At this base fee, a standard TIP-20 transfer (~50,000 gas) costs ~0.1 cent
+/// T1 base fee: 20 billion attodollars (2×10^10)
+///
+/// Attodollars are the atomic gas accounting units at 10^-18 USD precision.
+/// Basefee is denominated in attodollars.
+///
+/// At this basefee, a standard TIP-20 transfer (~50,000 gas) costs:
+/// - Gas: 50,000 × 20 billion attodollars/gas = 1 quadrillion attodollars
+/// - Tokens: 1 quadrillion attodollars / 10^12 = 1,000 microdollars
+/// - Economic: 1,000 microdollars = 0.001 USD = 0.1 cents
 pub const TEMPO_T1_BASE_FEE: u64 = 20_000_000_000;
 
 /// TIP-1010 general (non-payment) gas limit: 30 million gas per block.
@@ -63,6 +73,12 @@ pub struct TempoGenesisInfo {
     /// Activation timestamp for T1 hardfork.
     #[serde(skip_serializing_if = "Option::is_none")]
     t1_time: Option<u64>,
+    /// Activation timestamp for T1.A hardfork.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    t1a_time: Option<u64>,
+    /// Activation timestamp for T1.B hardfork.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    t1b_time: Option<u64>,
     /// Activation timestamp for T2 hardfork.
     #[serde(skip_serializing_if = "Option::is_none")]
     t2_time: Option<u64>,
@@ -88,6 +104,14 @@ impl TempoGenesisInfo {
 
     pub fn t1_time(&self) -> Option<u64> {
         self.t1_time
+    }
+
+    pub fn t1a_time(&self) -> Option<u64> {
+        self.t1a_time
+    }
+
+    pub fn t1b_time(&self) -> Option<u64> {
+        self.t1b_time
     }
 
     pub fn t2_time(&self) -> Option<u64> {
@@ -154,7 +178,7 @@ pub static PRESTO: LazyLock<Arc<TempoChainSpec>> = LazyLock::new(|| {
 
 /// Development chainspec with funded dev accounts and activated tempo hardforks
 ///
-/// `cargo x generate-genesis -o dev.json --accounts 10`
+/// `cargo x generate-genesis -o dev.json --accounts 10 --no-dkg-in-genesis`
 pub static DEV: LazyLock<Arc<TempoChainSpec>> = LazyLock::new(|| {
     let genesis: Genesis = serde_json::from_str(include_str!("./genesis/dev.json"))
         .expect("`./genesis/dev.json` must be present and deserializable");
@@ -183,6 +207,8 @@ impl TempoChainSpec {
         let info @ TempoGenesisInfo {
             t0_time,
             t1_time,
+            t1a_time,
+            t1b_time,
             t2_time,
             ..
         } = TempoGenesisInfo::extract_from(&genesis);
@@ -194,6 +220,8 @@ impl TempoChainSpec {
             (TempoHardfork::Genesis, Some(0)),
             (TempoHardfork::T0, t0_time),
             (TempoHardfork::T1, t1_time),
+            (TempoHardfork::T1A, t1a_time),
+            (TempoHardfork::T1B, t1b_time),
             (TempoHardfork::T2, t2_time),
         ]
         .into_iter()
@@ -216,6 +244,11 @@ impl TempoChainSpec {
     pub fn with_default_follow_url(mut self, url: &'static str) -> Self {
         self.default_follow_url = Some(url);
         self
+    }
+
+    /// Returns the mainnet chainspec.
+    pub fn mainnet() -> Self {
+        PRESTO.as_ref().clone()
     }
 }
 
@@ -411,18 +444,37 @@ mod tests {
             TempoHardfork::T0
         );
 
-        // At and after T1 activation
+        // At and after T1/T1A activation (both activate at 1770908400)
+        assert!(mainnet_chainspec.is_t1_active_at_timestamp(1770908400));
+        assert!(mainnet_chainspec.is_t1a_active_at_timestamp(1770908400));
         assert_eq!(
             mainnet_chainspec.tempo_hardfork_at(1770908400),
-            TempoHardfork::T1
+            TempoHardfork::T1A
         );
         assert_eq!(
             mainnet_chainspec.tempo_hardfork_at(1770908401),
-            TempoHardfork::T1
+            TempoHardfork::T1A
         );
+
+        // Before T1B activation (1771858800 = Feb 23rd 2026 16:00 CET)
+        assert!(!mainnet_chainspec.is_t1b_active_at_timestamp(1771858799));
+        assert_eq!(
+            mainnet_chainspec.tempo_hardfork_at(1771858799),
+            TempoHardfork::T1A
+        );
+
+        // At and after T1B activation
+        assert!(mainnet_chainspec.is_t1b_active_at_timestamp(1771858800));
+        assert_eq!(
+            mainnet_chainspec.tempo_hardfork_at(1771858800),
+            TempoHardfork::T1B
+        );
+
+        // T2 not yet scheduled on mainnet
+        assert!(!mainnet_chainspec.is_t2_active_at_timestamp(u64::MAX));
         assert_eq!(
             mainnet_chainspec.tempo_hardfork_at(u64::MAX),
-            TempoHardfork::T1
+            TempoHardfork::T1B
         );
 
         let moderato_genesis = super::TempoChainSpecParser::parse("moderato")
@@ -447,15 +499,32 @@ mod tests {
             moderato_genesis.tempo_hardfork_at(1770303601),
             TempoHardfork::T1
         );
+
+        // Before T1A/T1B activation (1771858800 = Feb 23rd 2026 16:00 CET)
         assert_eq!(
-            moderato_genesis.tempo_hardfork_at(u64::MAX),
+            moderato_genesis.tempo_hardfork_at(1771858799),
             TempoHardfork::T1
         );
 
-        let testnet_chainspec = super::TempoChainSpecParser::parse("testnet")
-            .expect("the mainnet chainspec must always be well formed");
+        // At and after T1A/T1B activation (both activate at 1771858800)
+        assert!(moderato_genesis.is_t1a_active_at_timestamp(1771858800));
+        assert!(moderato_genesis.is_t1b_active_at_timestamp(1771858800));
+        assert_eq!(
+            moderato_genesis.tempo_hardfork_at(1771858800),
+            TempoHardfork::T1B
+        );
 
-        // Should always return Genesis
+        // T2 not yet scheduled on moderato
+        assert!(!moderato_genesis.is_t2_active_at_timestamp(u64::MAX));
+        assert_eq!(
+            moderato_genesis.tempo_hardfork_at(u64::MAX),
+            TempoHardfork::T1B
+        );
+
+        let testnet_chainspec = super::TempoChainSpecParser::parse("testnet")
+            .expect("the testnet chainspec must always be well formed");
+
+        // Should always return Genesis (no hardfork timestamps on andantino)
         assert_eq!(
             testnet_chainspec.tempo_hardfork_at(0),
             TempoHardfork::Genesis
@@ -486,6 +555,8 @@ mod tests {
                     "chainId": 1234,
                     "t0Time": 0,
                     "t1Time": 0,
+                    "t1aTime": 0,
+                    "t1bTime": 0,
                     "t2Time": 0
                 },
                 "alloc": {}
@@ -499,6 +570,10 @@ mod tests {
         assert!(chainspec.is_t0_active_at_timestamp(1000));
         assert!(chainspec.is_t1_active_at_timestamp(0));
         assert!(chainspec.is_t1_active_at_timestamp(1000));
+        assert!(chainspec.is_t1a_active_at_timestamp(0));
+        assert!(chainspec.is_t1a_active_at_timestamp(1000));
+        assert!(chainspec.is_t1b_active_at_timestamp(0));
+        assert!(chainspec.is_t1b_active_at_timestamp(1000));
         assert!(chainspec.is_t2_active_at_timestamp(0));
         assert!(chainspec.is_t2_active_at_timestamp(1000));
 
