@@ -22,8 +22,8 @@ use commonware_cryptography::{
 use commonware_p2p::{AddressableManager, Blocker, Receiver, Sender};
 use commonware_parallel::Sequential;
 use commonware_runtime::{
-    Clock, ContextCell, Handle, Metrics, Network, Pacer, Spawner, Storage, buffer::paged::CacheRef,
-    spawn_cell,
+    BufferPooler, Clock, ContextCell, Handle, Metrics, Network, Pacer, Spawner, Storage,
+    buffer::paged::CacheRef, spawn_cell,
 };
 use commonware_storage::archive::immutable;
 use commonware_utils::NZU64;
@@ -38,7 +38,7 @@ use crate::{
     consensus::application,
     dkg,
     epoch::{self, SchemeProvider},
-    subblocks,
+    peer_manager, subblocks,
 };
 
 use super::block::Block;
@@ -97,7 +97,7 @@ pub struct Builder<TBlocker, TPeerManager> {
 
 impl<TBlocker, TPeerManager> Builder<TBlocker, TPeerManager>
 where
-    TBlocker: Blocker<PublicKey = PublicKey>,
+    TBlocker: Blocker<PublicKey = PublicKey> + Sync,
     TPeerManager: AddressableManager<PublicKey = PublicKey> + Sync,
 {
     pub fn with_execution_node(mut self, execution_node: TempoFullNode) -> Self {
@@ -108,7 +108,7 @@ where
     pub async fn try_init<TContext>(
         self,
         context: TContext,
-    ) -> eyre::Result<Engine<TBlocker, TContext, TPeerManager>>
+    ) -> eyre::Result<Engine<TContext, TBlocker, TPeerManager>>
     where
         TContext: Clock
             + governor::clock::Clock
@@ -118,7 +118,8 @@ where
             + Spawner
             + Storage
             + Metrics
-            + Network,
+            + Network
+            + BufferPooler,
     {
         let execution_node = self
             .execution_node
@@ -130,6 +131,8 @@ where
             .info
             .epoch_length()
             .ok_or_eyre("chainspec did not contain epochLength; cannot go on without it")?;
+
+        let epoch_strategy = FixedEpocher::new(NZU64!(epoch_length));
 
         info!(
             identity = %self.signer.public_key(),
@@ -147,22 +150,9 @@ where
             },
         );
 
-        let page_cache_ref = CacheRef::new(BUFFER_POOL_PAGE_SIZE, BUFFER_POOL_CAPACITY);
+        let page_cache_ref =
+            CacheRef::from_pooler(&context, BUFFER_POOL_PAGE_SIZE, BUFFER_POOL_CAPACITY);
 
-        // XXX: All hard-coded values here are the same as prior to commonware
-        // making the resolver configurable in
-        // https://github.com/commonwarexyz/monorepo/commit/92870f39b4a9e64a28434b3729ebff5aba67fb4e
-        let resolver_config = commonware_consensus::marshal::resolver::p2p::Config {
-            public_key: self.signer.public_key(),
-            provider: self.peer_manager.clone(),
-            mailbox_size: self.mailbox_size,
-            blocker: self.blocker.clone(),
-            initial: Duration::from_secs(1),
-            timeout: Duration::from_secs(2),
-            fetch_retry_timeout: Duration::from_millis(100),
-            priority_requests: false,
-            priority_responses: false,
-        };
         let scheme_provider = SchemeProvider::new();
 
         const FINALIZATIONS_BY_HEIGHT: &str = "finalizations-by-height";
@@ -261,7 +251,6 @@ where
         .wrap_err("failed to initialize finalizations by height archive")?;
         info!(elapsed = ?start.elapsed(), "restored finalizations by height archive");
 
-        let epoch_strategy = FixedEpocher::new(NZU64!(epoch_length));
         // TODO(janis): forward `last_finalized_height` to application so it can
         // forward missing blocks to EL.
         let (marshal, marshal_mailbox, last_finalized_height) = marshal::Actor::init(
@@ -291,6 +280,31 @@ where
             },
         )
         .await;
+
+        let (peer_manager, peer_manager_mailbox) = peer_manager::init(
+            context.with_label("peer_manager"),
+            peer_manager::Config {
+                execution_node: execution_node.clone(),
+                oracle: self.peer_manager.clone(),
+                epoch_strategy: epoch_strategy.clone(),
+                last_finalized_height,
+            },
+        );
+
+        // XXX: All hard-coded values here are the same as prior to commonware
+        // making the resolver configurable in
+        // https://github.com/commonwarexyz/monorepo/commit/92870f39b4a9e64a28434b3729ebff5aba67fb4e
+        let resolver_config = commonware_consensus::marshal::resolver::p2p::Config {
+            public_key: self.signer.public_key(),
+            provider: peer_manager_mailbox.clone(),
+            mailbox_size: self.mailbox_size,
+            blocker: self.blocker.clone(),
+            initial: Duration::from_secs(1),
+            timeout: Duration::from_secs(2),
+            fetch_retry_timeout: Duration::from_millis(100),
+            priority_requests: false,
+            priority_responses: false,
+        };
 
         let subblocks = subblocks::Actor::new(subblocks::Config {
             context: context.clone(),
@@ -370,7 +384,6 @@ where
                 namespace: crate::config::NAMESPACE.to_vec(),
                 me: self.signer.clone(),
                 partition_prefix: format!("{}_dkg_manager", self.partition_prefix),
-                peer_manager: self.peer_manager.clone(),
             },
         )
         .await
@@ -396,6 +409,9 @@ where
             epoch_manager,
             epoch_manager_mailbox,
 
+            peer_manager,
+            peer_manager_mailbox,
+
             feed,
 
             subblocks,
@@ -403,10 +419,10 @@ where
     }
 }
 
-pub struct Engine<TBlocker, TContext, TPeerManager>
+pub struct Engine<TContext, TBlocker, TPeerManager>
 where
-    TBlocker: Blocker<PublicKey = PublicKey>,
-    TContext: Clock
+    TContext: BufferPooler
+        + Clock
         + governor::clock::Clock
         + Rng
         + CryptoRng
@@ -415,6 +431,7 @@ where
         + Pacer
         + Spawner
         + Storage,
+    TBlocker: Blocker<PublicKey = PublicKey>,
     TPeerManager: AddressableManager<PublicKey = PublicKey>,
 {
     context: ContextCell<TContext>,
@@ -424,7 +441,7 @@ where
     broadcast: buffered::Engine<TContext, PublicKey, Block>,
     broadcast_mailbox: buffered::Mailbox<PublicKey, Block>,
 
-    dkg_manager: dkg::manager::Actor<TContext, TPeerManager>,
+    dkg_manager: dkg::manager::Actor<TContext>,
     dkg_manager_mailbox: dkg::manager::Mailbox,
 
     /// Acts as the glue between the consensus and execution layers implementing
@@ -438,24 +455,27 @@ where
     executor_mailbox: crate::executor::Mailbox,
 
     /// Resolver config that will be passed to the marshal actor upon start.
-    resolver_config: marshal::resolver::p2p::Config<PublicKey, TPeerManager, TBlocker>,
+    resolver_config: marshal::resolver::p2p::Config<PublicKey, peer_manager::Mailbox, TBlocker>,
 
     /// Listens to consensus events and syncs blocks from the network to the
     /// local node.
     marshal: crate::alias::marshal::Actor<TContext>,
 
-    epoch_manager: epoch::manager::Actor<TBlocker, TContext>,
+    epoch_manager: epoch::manager::Actor<TContext, TBlocker>,
     epoch_manager_mailbox: epoch::manager::Mailbox,
+
+    peer_manager: peer_manager::Actor<TContext, TPeerManager>,
+    peer_manager_mailbox: peer_manager::Mailbox,
 
     feed: crate::feed::Actor<TContext>,
 
     subblocks: subblocks::Actor<TContext>,
 }
 
-impl<TBlocker, TContext, TPeerManager> Engine<TBlocker, TContext, TPeerManager>
+impl<TContext, TBlocker, TPeerManager> Engine<TContext, TBlocker, TPeerManager>
 where
-    TBlocker: Blocker<PublicKey = PublicKey>,
-    TContext: Clock
+    TContext: BufferPooler
+        + Clock
         + governor::clock::Clock
         + Rng
         + CryptoRng
@@ -464,6 +484,7 @@ where
         + Pacer
         + Spawner
         + Storage,
+    TBlocker: Blocker<PublicKey = PublicKey> + Sync,
     TPeerManager: AddressableManager<PublicKey = PublicKey> + Sync,
 {
     #[expect(
@@ -551,6 +572,8 @@ where
             impl Receiver<PublicKey = PublicKey>,
         ),
     ) -> eyre::Result<()> {
+        let peer_manager = self.peer_manager.start();
+
         let broadcast = self.broadcast.start(broadcast_channel);
         let resolver =
             marshal::resolver::p2p::init(&self.context, self.resolver_config, marshal_channel);
@@ -561,7 +584,10 @@ where
         let marshal = self.marshal.start(
             Reporters::from((
                 self.epoch_manager_mailbox,
-                Reporters::from((self.executor_mailbox, self.dkg_manager_mailbox.clone())),
+                Reporters::from((
+                    self.executor_mailbox,
+                    Reporters::from((self.dkg_manager_mailbox.clone(), self.peer_manager_mailbox)),
+                )),
             )),
             self.broadcast_mailbox,
             resolver,
@@ -587,6 +613,7 @@ where
             feed,
             marshal,
             dkg_manager,
+            peer_manager,
             subblocks,
         ])
         .await
