@@ -78,7 +78,7 @@ fn subblocks_are_included() {
                 ConsensusEngineEvent::CanonicalBlockAdded(block, _) => block,
             };
 
-            let receipts = &block.execution_outcome().receipts()[0];
+            let receipts = &block.execution_outcome().receipts;
 
             // Assert that block only contains our subblock transactions and the system transactions
             assert_eq!(
@@ -107,8 +107,9 @@ fn subblocks_are_included() {
             if !expected_transactions.is_empty() {
                 let fee_token_storage = &block
                     .execution_outcome()
-                    .state()
-                    .account(&DEFAULT_FEE_TOKEN)
+                    .state
+                    .state
+                    .get(&DEFAULT_FEE_TOKEN)
                     .unwrap()
                     .storage;
 
@@ -184,7 +185,7 @@ fn subblocks_are_included_with_failing_txs() {
                 ConsensusEngineEvent::InvalidBlock(_) => unreachable!("unexpected invalid block"),
                 ConsensusEngineEvent::CanonicalBlockAdded(block, _) => block,
             };
-            let receipts = &block.execution_outcome().receipts()[0];
+            let receipts = &block.execution_outcome().receipts;
 
             // Assert that block only contains our subblock transactions and system transactions
             assert_eq!(
@@ -258,8 +259,9 @@ fn subblocks_are_included_with_failing_txs() {
 
                 let slot = block
                     .execution_outcome()
-                    .state()
-                    .account(&NONCE_PRECOMPILE_ADDRESS)
+                    .state
+                    .state
+                    .get(&NONCE_PRECOMPILE_ADDRESS)
                     .unwrap()
                     .storage
                     .get(&nonce_slot)
@@ -275,8 +277,9 @@ fn subblocks_are_included_with_failing_txs() {
             for (fee_recipient, expected_fee) in expected_fees {
                 let fee_token_storage = &block
                     .execution_outcome()
-                    .state()
-                    .account(&DEFAULT_FEE_TOKEN)
+                    .state
+                    .state
+                    .get(&DEFAULT_FEE_TOKEN)
                     .unwrap()
                     .storage;
 
