@@ -1,13 +1,13 @@
 use crate::{TempoBlockExecutionCtx, evm::TempoEvm};
 use alloy_consensus::{Transaction, transaction::TxHashRef};
 use alloy_evm::{
-    Database, Evm,
+    Database, Evm, RecoveredTx,
     block::{
         BlockExecutionError, BlockExecutionResult, BlockExecutor, BlockValidationError,
-        ExecutableTx, OnStateHook,
+        ExecutableTx, OnStateHook, TxResult,
     },
     eth::{
-        EthBlockExecutor,
+        EthBlockExecutor, EthTxResult,
         receipt_builder::{ReceiptBuilder, ReceiptBuilderCtx},
     },
 };
@@ -18,11 +18,18 @@ use commonware_cryptography::{
     Verifier,
     ed25519::{PublicKey, Signature},
 };
-use reth_revm::{Inspector, State, context::result::ResultAndState};
+use reth_revm::{
+    DatabaseCommit, Inspector, State,
+    context::result::ResultAndState,
+    context_interface::JournalTr,
+    state::{Account, Bytecode, EvmState},
+};
 use std::collections::{HashMap, HashSet};
-use tempo_chainspec::TempoChainSpec;
+use tempo_chainspec::{TempoChainSpec, hardfork::TempoHardforks};
+use tempo_contracts::precompiles::VALIDATOR_CONFIG_V2_ADDRESS;
 use tempo_primitives::{
-    SubBlock, SubBlockMetadata, TempoReceipt, TempoTxEnvelope, subblock::PartialValidatorKey,
+    SubBlock, SubBlockMetadata, TempoReceipt, TempoTxEnvelope, TempoTxType,
+    subblock::PartialValidatorKey,
 };
 use tempo_revm::{TempoHaltReason, evm::TempoContext};
 use tracing::trace;
@@ -46,30 +53,53 @@ pub(crate) enum BlockSection {
 /// Builder for [`TempoReceipt`].
 #[derive(Debug, Clone, Copy, Default)]
 #[non_exhaustive]
-pub(crate) struct TempoReceiptBuilder;
+pub struct TempoReceiptBuilder;
 
 impl ReceiptBuilder for TempoReceiptBuilder {
     type Transaction = TempoTxEnvelope;
     type Receipt = TempoReceipt;
 
-    fn build_receipt<E: Evm>(
-        &self,
-        ctx: ReceiptBuilderCtx<'_, Self::Transaction, E>,
-    ) -> Self::Receipt {
+    fn build_receipt<E: Evm>(&self, ctx: ReceiptBuilderCtx<'_, TempoTxType, E>) -> Self::Receipt {
         let ReceiptBuilderCtx {
-            tx,
+            tx_type,
             result,
             cumulative_gas_used,
             ..
         } = ctx;
         TempoReceipt {
-            tx_type: tx.tx_type(),
+            tx_type,
             // Success flag was added in `EIP-658: Embedding transaction status code in
             // receipts`.
             success: result.is_success(),
             cumulative_gas_used,
             logs: result.into_logs(),
         }
+    }
+}
+
+/// The result of executing a Tempo transaction.
+///
+/// This is an extension of [`EthTxResult`] with context necessary for committing a Tempo transaction.
+#[derive(Debug)]
+pub(crate) struct TempoTxResult<H> {
+    /// Inner transaction execution result.
+    inner: EthTxResult<H, TempoTxType>,
+    /// Next section of the block.
+    next_section: BlockSection,
+    /// Whether the transaction is a payment transaction.
+    is_payment: bool,
+    /// Full transaction that is being committed.
+    ///
+    /// This is only populated for subblock transactions for which we need to store
+    /// the full transaction encoding for later validation of subblock hash.
+    tx: Option<TempoTxEnvelope>,
+}
+
+impl<H> TxResult for TempoTxResult<H> {
+    type HaltReason = H;
+
+    fn result(&self) -> &ResultAndState<Self::HaltReason> {
+        self.inner.result()
     }
 }
 
@@ -333,9 +363,33 @@ where
     type Transaction = TempoTxEnvelope;
     type Receipt = TempoReceipt;
     type Evm = TempoEvm<&'a mut State<DB>, I>;
+    type Result = TempoTxResult<TempoHaltReason>;
 
     fn apply_pre_execution_changes(&mut self) -> Result<(), alloy_evm::block::BlockExecutionError> {
-        self.inner.apply_pre_execution_changes()
+        self.inner.apply_pre_execution_changes()?;
+
+        // Deploy 0xEF marker bytecode to ValidatorConfigV2 when T2 activates.
+        let timestamp = self.evm().block().timestamp.to::<u64>();
+        if self.inner.spec.is_t2_active_at_timestamp(timestamp) {
+            let db = self.evm_mut().ctx_mut().journaled_state.db_mut();
+            let acc = db
+                .load_cache_account(VALIDATOR_CONFIG_V2_ADDRESS)
+                .map_err(BlockExecutionError::other)?;
+            let mut info = acc.account_info().unwrap_or_default();
+            if info.is_empty_code_hash() {
+                let code = Bytecode::new_legacy([0xef].into());
+                info.code_hash = code.hash_slow();
+                info.code = Some(code);
+                let mut account: Account = info.into();
+                account.mark_touch();
+                db.commit(EvmState::from_iter([(
+                    VALIDATOR_CONFIG_V2_ADDRESS,
+                    account,
+                )]));
+            }
+        }
+
+        Ok(())
     }
 
     fn receipts(&self) -> &[Self::Receipt] {
@@ -345,10 +399,12 @@ where
     fn execute_transaction_without_commit(
         &mut self,
         tx: impl ExecutableTx<Self>,
-    ) -> Result<ResultAndState<TempoHaltReason>, BlockExecutionError> {
+    ) -> Result<Self::Result, BlockExecutionError> {
+        let (tx_env, recovered) = tx.into_parts();
+
         let beneficiary = self.evm_mut().ctx_mut().block.beneficiary;
         // If we are dealing with a subblock transaction, configure the fee recipient context.
-        if let Some(validator) = tx.tx().subblock_proposer() {
+        if let Some(validator) = recovered.tx().subblock_proposer() {
             let fee_recipient = *self
                 .subblock_fee_recipients
                 .get(&validator)
@@ -356,21 +412,33 @@ where
 
             self.evm_mut().ctx_mut().block.beneficiary = fee_recipient;
         }
-        let result = self.inner.execute_transaction_without_commit(tx);
+        let result = self
+            .inner
+            .execute_transaction_without_commit((tx_env, &recovered));
 
         self.evm_mut().ctx_mut().block.beneficiary = beneficiary;
 
-        result
+        let inner = result?;
+
+        let next_section = self.validate_tx(recovered.tx(), inner.result.result.gas_used())?;
+        Ok(TempoTxResult {
+            inner,
+            next_section,
+            is_payment: recovered.tx().is_payment(),
+            tx: matches!(next_section, BlockSection::SubBlock { .. })
+                .then(|| recovered.tx().clone()),
+        })
     }
 
-    fn commit_transaction(
-        &mut self,
-        output: ResultAndState<TempoHaltReason>,
-        tx: impl ExecutableTx<Self>,
-    ) -> Result<u64, BlockExecutionError> {
-        let next_section = self.validate_tx(tx.tx(), output.result.gas_used())?;
+    fn commit_transaction(&mut self, output: Self::Result) -> Result<u64, BlockExecutionError> {
+        let TempoTxResult {
+            inner,
+            next_section,
+            is_payment,
+            tx,
+        } = output;
 
-        let gas_used = self.inner.commit_transaction(output, &tx)?;
+        let gas_used = self.inner.commit_transaction(inner)?;
 
         // TODO: remove once revm supports emitting logs for reverted transactions
         //
@@ -393,12 +461,11 @@ where
             }
             BlockSection::NonShared => {
                 self.non_shared_gas_left -= gas_used;
-                if !tx.tx().is_payment() {
+                if !is_payment {
                     self.non_payment_gas_left -= gas_used;
                 }
             }
             BlockSection::SubBlock { proposer } => {
-                // record subblock transactions to verify later
                 let last_subblock = if let Some(last) = self
                     .seen_subblocks
                     .last_mut()
@@ -410,7 +477,9 @@ where
                     self.seen_subblocks.last_mut().unwrap()
                 };
 
-                last_subblock.1.push(tx.tx().clone());
+                last_subblock.1.push(tx.ok_or_else(|| {
+                    BlockExecutionError::msg("missing tx for subblock transaction")
+                })?);
             }
             BlockSection::GasIncentive => {
                 self.incentive_gas_used += gas_used;
@@ -426,7 +495,6 @@ where
     fn finish(
         self,
     ) -> Result<(Self::Evm, BlockExecutionResult<Self::Receipt>), BlockExecutionError> {
-        // Check that we ended in the System section with all end-of-block system txs seen
         if self.section
             != (BlockSection::System {
                 seen_subblocks_signatures: true,
@@ -488,7 +556,7 @@ where
 mod tests {
     use super::*;
     use crate::test_utils::{TestExecutorBuilder, test_chainspec, test_evm};
-    use alloy_consensus::{Signed, TxLegacy, transaction::Recovered};
+    use alloy_consensus::{Signed, TxLegacy};
     use alloy_evm::{block::BlockExecutor, eth::receipt_builder::ReceiptBuilder};
     use alloy_primitives::{Bytes, Log, Signature, TxKind, bytes::BytesMut};
     use alloy_rlp::Encodable;
@@ -538,7 +606,7 @@ mod tests {
         let cumulative_gas_used = 21000;
 
         let receipt = builder.build_receipt(ReceiptBuilderCtx {
-            tx: &tx,
+            tx_type: tx.tx_type(),
             evm: &evm,
             result,
             state: &Default::default(),
@@ -1033,21 +1101,28 @@ mod tests {
         // Apply pre-execution changes first
         executor.apply_pre_execution_changes().unwrap();
 
-        // Create execution result
-        let output = ResultAndState {
-            result: revm::context::result::ExecutionResult::Success {
-                reason: revm::context::result::SuccessReason::Return,
-                gas_used: 21000,
-                gas_refunded: 0,
-                logs: vec![],
-                output: revm::context::result::Output::Call(Bytes::new()),
+        let tx = create_legacy_tx();
+        let output = TempoTxResult {
+            inner: EthTxResult {
+                result: ResultAndState {
+                    result: revm::context::result::ExecutionResult::Success {
+                        reason: revm::context::result::SuccessReason::Return,
+                        gas_used: 21000,
+                        gas_refunded: 0,
+                        logs: vec![],
+                        output: revm::context::result::Output::Call(Bytes::new()),
+                    },
+                    state: Default::default(),
+                },
+                blob_gas_used: 0,
+                tx_type: tx.tx_type(),
             },
-            state: Default::default(),
+            next_section: BlockSection::NonShared,
+            is_payment: false,
+            tx: None,
         };
 
-        let tx = create_legacy_tx();
-        let recovered_tx = Recovered::new_unchecked(tx, Address::ZERO);
-        let gas_used = executor.commit_transaction(output, &recovered_tx).unwrap();
+        let gas_used = executor.commit_transaction(output).unwrap();
 
         assert_eq!(gas_used, 21000);
         assert_eq!(executor.section(), BlockSection::NonShared);
@@ -1080,5 +1155,24 @@ mod tests {
         assert!(result.is_err());
         let err = result.err().unwrap();
         assert_eq!(err.to_string(), "end-of-block system transactions not seen");
+    }
+
+    #[test]
+    fn test_apply_pre_execution_deploys_validator_v2_code() {
+        use std::sync::Arc;
+        use tempo_chainspec::spec::DEV;
+
+        // Dev chainspec has t2Time: 0, so T2 is active at any timestamp.
+        let chainspec = Arc::new(TempoChainSpec::from_genesis(DEV.genesis().clone()));
+        let mut db = State::builder().with_bundle_update().build();
+        let mut executor = TestExecutorBuilder::default()
+            .with_parent_beacon_block_root(B256::ZERO)
+            .build(&mut db, &chainspec);
+
+        executor.apply_pre_execution_changes().unwrap();
+
+        let acc = db.load_cache_account(VALIDATOR_CONFIG_V2_ADDRESS).unwrap();
+        let info = acc.account_info().unwrap();
+        assert!(!info.is_empty_code_hash());
     }
 }
