@@ -3,6 +3,7 @@ use std::collections::HashMap;
 use alloy_evm::eth::EthBlockExecutionCtx;
 use alloy_primitives::{Address, B256};
 use reth_evm::NextBlockEnvAttributes;
+use tempo_contracts::precompiles::TIP_FEE_MANAGER_ADDRESS;
 use tempo_primitives::subblock::PartialValidatorKey;
 
 /// Execution context for Tempo block.
@@ -53,11 +54,12 @@ impl reth_rpc_eth_api::helpers::pending_block::BuildPendingEnv<tempo_primitives:
         // (PendingBlockKind::None) - blocks require consensus data that RPC doesn't have.
         let mut inner = NextBlockEnvAttributes::build_pending_env(parent);
 
-        // Use Address::ZERO as fee recipient so the EVM resolves the default fee token
-        // (PathUSD) for RPC simulation calls (eth_call, eth_estimateGas). Without this,
-        // the actual block producer's fee token is used, which may have no AMM pool with
-        // the user's fee token, causing spurious "insufficient liquidity" errors.
-        inner.suggested_fee_recipient = alloy_primitives::Address::ZERO;
+        // Use TIP_FEE_MANAGER_ADDRESS as a sentinel fee recipient so the EVM resolves the
+        // default fee token (PathUSD) for RPC simulations (eth_call, eth_estimateGas).
+        // This address can never be the sender of a transaction, so its validatorTokens
+        // mapping is guaranteed to be zero, which falls back to DEFAULT_FEE_TOKEN.
+        // NOTE: Address::ZERO cannot be used because genesis maps it to the "DONOTUSE" token.
+        inner.suggested_fee_recipient = TIP_FEE_MANAGER_ADDRESS;
 
         Self {
             inner,
@@ -104,9 +106,8 @@ mod tests {
         assert_eq!(pending_env.timestamp_millis_part, timestamp_millis_part);
         assert!(pending_env.subblock_fee_recipients.is_empty());
         assert_eq!(
-            pending_env.inner.suggested_fee_recipient,
-            Address::ZERO,
-            "pending env uses Address::ZERO so RPC resolves default fee token"
+            pending_env.inner.suggested_fee_recipient, TIP_FEE_MANAGER_ADDRESS,
+            "pending env uses TIP_FEE_MANAGER_ADDRESS so RPC resolves default fee token"
         );
     }
 }
