@@ -17,7 +17,7 @@ use alloy::{
 use alloy_primitives::B256;
 use alloy_rpc_types_engine::PayloadAttributes;
 use reth_e2e_test_utils::setup;
-use reth_ethereum::tasks::TaskManager;
+use reth_ethereum::tasks::Runtime;
 use reth_node_api::{FullNodeComponents, PayloadBuilderAttributes};
 use reth_node_builder::{NodeBuilder, NodeConfig, NodeHandle, rpc::RethRpcAddOns};
 use reth_node_core::args::RpcServerArgs;
@@ -80,8 +80,8 @@ pub(crate) enum NodeSource {
     LocalNode(String),
 }
 
-/// Type alias for a local test node and task manager
-pub(crate) type LocalTestNode = (Box<dyn TestNodeHandle>, TaskManager);
+/// Type alias for a local test node and runtime
+pub(crate) type LocalTestNode = (Box<dyn TestNodeHandle>, Runtime);
 
 /// Trait wrapper around NodeHandle to simplify function return types
 pub(crate) trait TestNodeHandle: Send {}
@@ -137,16 +137,12 @@ pub(crate) async fn await_receipts(
 pub(crate) struct SingleNodeSetup {
     /// The node handle for direct manipulation (inject_tx, advance_block, etc.)
     pub node: reth_e2e_test_utils::NodeHelperType<TempoNode>,
-    /// Task manager that must be kept alive for the node to function
-    _tasks: TaskManager,
 }
 
 /// Result type for multi-node setup
 pub(crate) struct MultiNodeSetup {
     /// Node handles for direct manipulation
     pub nodes: Vec<reth_e2e_test_utils::NodeHelperType<TempoNode>>,
-    /// Task manager that must be kept alive for nodes to function
-    _tasks: TaskManager,
 }
 
 /// Result type for HTTP-only setup (no direct node access)
@@ -231,7 +227,7 @@ impl TestNodeBuilder {
 
         let chain_spec = self.build_chain_spec()?;
 
-        let (mut nodes, tasks, _wallet) = setup::<TempoNode>(
+        let (mut nodes, _wallet) = setup::<TempoNode>(
             1,
             Arc::new(chain_spec),
             self.is_dev,
@@ -241,10 +237,7 @@ impl TestNodeBuilder {
 
         let node = nodes.remove(0);
 
-        Ok(SingleNodeSetup {
-            node,
-            _tasks: tasks,
-        })
+        Ok(SingleNodeSetup { node })
     }
 
     /// Build multiple nodes with direct access
@@ -263,7 +256,7 @@ impl TestNodeBuilder {
 
         let chain_spec = self.build_chain_spec()?;
 
-        let (nodes, tasks, _wallet) = setup::<TempoNode>(
+        let (nodes, _wallet) = setup::<TempoNode>(
             self.node_count,
             Arc::new(chain_spec),
             self.is_dev,
@@ -271,10 +264,7 @@ impl TestNodeBuilder {
         )
         .await?;
 
-        Ok(MultiNodeSetup {
-            nodes,
-            _tasks: tasks,
-        })
+        Ok(MultiNodeSetup { nodes })
     }
 
     /// Build HTTP-only setup
@@ -286,7 +276,7 @@ impl TestNodeBuilder {
             });
         }
 
-        let tasks = TaskManager::current();
+        let runtime = Runtime::test_with_handle(tokio::runtime::Handle::current());
         let chain_spec = self.build_chain_spec()?;
         let static_validator = self
             .custom_validator
@@ -306,7 +296,7 @@ impl TestNodeBuilder {
         node_config.dev.block_time = Some(Duration::from_millis(100));
 
         let node_handle = NodeBuilder::new(node_config.clone())
-            .testing_node(tasks.executor())
+            .testing_node(runtime.clone())
             .node(TempoNode::default())
             .launch_with_debug_capabilities()
             .map_debug_payload_attributes(move |mut attributes| {
@@ -329,7 +319,7 @@ impl TestNodeBuilder {
 
         Ok(HttpOnlySetup {
             http_url,
-            local_node: Some((Box::new(node_handle), tasks)),
+            local_node: Some((Box::new(node_handle), runtime)),
         })
     }
 
