@@ -14,6 +14,13 @@ use std::{sync::Arc, time::Instant};
 /// If a token isn't unpaused within this time, we clear all pending transactions.
 pub const PAUSED_TX_TIMEOUT: std::time::Duration = std::time::Duration::from_secs(30 * 60); // 30 minutes
 
+/// Global cap on the total number of paused transactions across all tokens.
+///
+/// Without this cap, an attacker could repeatedly fill the main pool, trigger a pause, and shift
+/// transactions into the paused pool indefinitely. This bounds memory usage regardless of how many
+/// tokens are paused or how frequently pause events occur.
+pub const PAUSED_POOL_GLOBAL_CAP: usize = 10_000;
+
 /// Entry in the paused pool.
 #[derive(Debug, Clone)]
 pub struct PausedEntry {
@@ -63,10 +70,34 @@ impl PausedFeeTokenPool {
     ///
     /// Takes the full batch at once since all transactions for a token
     /// are paused together. The pause timestamp is recorded at insertion time.
-    pub fn insert_batch(&mut self, fee_token: Address, entries: Vec<PausedEntry>) {
+    ///
+    /// Enforces [`PAUSED_POOL_GLOBAL_CAP`]: if adding the batch would exceed the cap,
+    /// the oldest-paused tokens are evicted first to make room. If the batch itself
+    /// exceeds the cap, it is truncated.
+    ///
+    /// Returns the number of existing entries that were evicted to make room.
+    pub fn insert_batch(&mut self, fee_token: Address, entries: Vec<PausedEntry>) -> usize {
         if entries.is_empty() {
-            return;
+            return 0;
         }
+
+        let current = self.len();
+        let incoming = entries.len();
+        let available = PAUSED_POOL_GLOBAL_CAP.saturating_sub(current);
+        let mut evicted = 0;
+
+        if incoming > available {
+            let need = incoming - available;
+            evicted = self.evict_oldest(need);
+        }
+
+        let remaining_capacity = PAUSED_POOL_GLOBAL_CAP.saturating_sub(self.len());
+        let to_insert = if incoming > remaining_capacity {
+            entries.into_iter().take(remaining_capacity).collect()
+        } else {
+            entries
+        };
+
         self.by_token
             .entry(fee_token)
             .or_insert_with(|| PausedTokenMeta {
@@ -74,7 +105,32 @@ impl PausedFeeTokenPool {
                 entries: Vec::new(),
             })
             .entries
-            .extend(entries);
+            .extend(to_insert);
+
+        evicted
+    }
+
+    /// Evicts at least `need` entries from the oldest-paused tokens.
+    ///
+    /// Returns the total number of entries evicted.
+    fn evict_oldest(&mut self, need: usize) -> usize {
+        let mut tokens_by_age: Vec<_> = self
+            .by_token
+            .iter()
+            .map(|(addr, meta)| (*addr, meta.paused_at))
+            .collect();
+        tokens_by_age.sort_unstable_by_key(|(_, paused_at)| *paused_at);
+
+        let mut evicted = 0;
+        for (token, _) in tokens_by_age {
+            if evicted >= need {
+                break;
+            }
+            if let Some(meta) = self.by_token.remove(&token) {
+                evicted += meta.entries.len();
+            }
+        }
+        evicted
     }
 
     /// Drains all transactions for a given fee token.
@@ -137,15 +193,22 @@ impl PausedFeeTokenPool {
 
     /// Removes transactions matching invalidation criteria from the paused pool.
     ///
-    /// This handles both revoked keys and spending limit updates in a single pass.
+    /// This handles revoked keys, spending limit updates, and spending limit spends
+    /// in a single pass. The `spending_limit_spends` parameter captures (account, key_id,
+    /// fee_token) combos from keychain txs that were included in the block and decremented
+    /// limits via `verify_and_update_spending()`.
     /// Uses account-keyed indexes for O(1) account lookup per transaction.
     /// Returns the number of transactions removed.
     pub fn evict_invalidated(
         &mut self,
         revoked_keys: &RevokedKeys,
         spending_limit_updates: &SpendingLimitUpdates,
+        spending_limit_spends: &SpendingLimitUpdates,
     ) -> usize {
-        if revoked_keys.is_empty() && spending_limit_updates.is_empty() {
+        if revoked_keys.is_empty()
+            && spending_limit_updates.is_empty()
+            && spending_limit_spends.is_empty()
+        {
             return 0;
         }
 
@@ -158,6 +221,7 @@ impl PausedFeeTokenPool {
                 };
                 !subject.matches_revoked(revoked_keys)
                     && !subject.matches_spending_limit_update(spending_limit_updates)
+                    && !subject.matches_spending_limit_update(spending_limit_spends)
             });
             count += before - meta.entries.len();
         }
@@ -232,6 +296,100 @@ mod tests {
         let evicted = pool.evict_expired(150);
         assert_eq!(evicted, 1);
         assert_eq!(pool.len(), 2);
+    }
+
+    #[test]
+    fn test_global_cap_evicts_oldest() {
+        let mut pool = PausedFeeTokenPool::new();
+
+        let token_a = Address::random();
+        let token_b = Address::random();
+
+        let make_entries = |n: usize| -> Vec<PausedEntry> {
+            (0..n)
+                .map(|_| PausedEntry {
+                    tx: create_valid_tx(Address::random()),
+                    valid_before: None,
+                })
+                .collect()
+        };
+
+        // Fill to cap
+        let evicted = pool.insert_batch(token_a, make_entries(PAUSED_POOL_GLOBAL_CAP));
+        assert_eq!(evicted, 0);
+        assert_eq!(pool.len(), PAUSED_POOL_GLOBAL_CAP);
+
+        // Inserting more should evict token_a (oldest) to make room
+        let evicted = pool.insert_batch(token_b, make_entries(100));
+        assert!(evicted > 0);
+        assert!(pool.len() <= PAUSED_POOL_GLOBAL_CAP);
+        assert_eq!(pool.count_for_token(&token_b), 100);
+    }
+
+    #[test]
+    fn test_global_cap_truncates_oversized_batch() {
+        let mut pool = PausedFeeTokenPool::new();
+        let token = Address::random();
+
+        let entries: Vec<_> = (0..PAUSED_POOL_GLOBAL_CAP + 500)
+            .map(|_| PausedEntry {
+                tx: create_valid_tx(Address::random()),
+                valid_before: None,
+            })
+            .collect();
+
+        let evicted = pool.insert_batch(token, entries);
+        assert_eq!(evicted, 0);
+        assert_eq!(pool.len(), PAUSED_POOL_GLOBAL_CAP);
+    }
+
+    #[test]
+    fn test_evict_invalidated_with_spending_limit_spends() {
+        let mut pool = PausedFeeTokenPool::new();
+        let user_address = Address::random();
+        let fee_token = Address::random();
+
+        // Create a keychain-signed tx using the builder
+        let access_key_signer = alloy_signer_local::PrivateKeySigner::random();
+        let key_id = alloy_signer::Signer::address(&access_key_signer);
+        let tx = TxBuilder::aa(user_address)
+            .fee_token(fee_token)
+            .build_keychain(user_address, &access_key_signer);
+        let tx = Arc::new(wrap_valid_tx(
+            tx,
+            reth_transaction_pool::TransactionOrigin::External,
+        ));
+
+        // Also add a non-keychain tx that should NOT be evicted
+        let other_tx = create_valid_tx(Address::random());
+
+        pool.insert_batch(
+            fee_token,
+            vec![
+                PausedEntry {
+                    tx,
+                    valid_before: None,
+                },
+                PausedEntry {
+                    tx: other_tx,
+                    valid_before: None,
+                },
+            ],
+        );
+        assert_eq!(pool.len(), 2);
+
+        // Build spending_limit_spends matching the keychain tx
+        let mut spends = SpendingLimitUpdates::new();
+        spends.insert(user_address, key_id, fee_token);
+
+        let evicted =
+            pool.evict_invalidated(&RevokedKeys::new(), &SpendingLimitUpdates::new(), &spends);
+
+        assert_eq!(
+            evicted, 1,
+            "Should evict the keychain tx matching the spend"
+        );
+        assert_eq!(pool.len(), 1, "Non-keychain tx should remain");
     }
 
     #[test]
